Updated vec3 to dvec3 for the new relative to eye camera. Introduced a new bug, voxels are inverted and camera is inverted will address in the next commit.

+++ src/scene/Scene.hpp
@@ -84,7 +84,7 @@ namespace Crescendo {
                 entities.erase(entities.begin() + index); 
 
                 // Re-sync
-                for (int i = 0; i < entities.size(); i++) {
+                for (size_t i = 0; i < entities.size(); i++) {
                     entities[i]->index = i;
                 }
             }
